+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Types of changes are:
 
 ## [Unreleased]
 
+## [0.20.0] - 2022-11-16
+
 ### Features
 
 - Plugin support. Now commands can also be loaded from installed packages.
@@ -289,7 +291,8 @@ Commands can raise `AssertionError` exceptions to tell `delfino` some pre-condit
 
 - Initial copy of source codes.
 
-[Unreleased]: https://github.com/radeklat/settings-doc/compare/0.19.1...HEAD
+[Unreleased]: https://github.com/radeklat/settings-doc/compare/0.20.0...HEAD
+[0.20.0]: https://github.com/radeklat/settings-doc/compare/0.19.1...0.20.0
 [0.19.1]: https://github.com/radeklat/settings-doc/compare/0.19.0...0.19.1
 [0.19.0]: https://github.com/radeklat/settings-doc/compare/0.18.2...0.19.0
 [0.18.2]: https://github.com/radeklat/settings-doc/compare/0.18.1...0.18.2